backend/pyproject.toml
[project]
name = "interactive-learning-backend"
version = "0.1.0"
description = "AI-powered interactive learning platform backend"
requires-python = ">=3.9"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # LangChain / LangGraph
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",

    # Document processing
    "langchain-text-splitters>=0.3.0",
    "pypdf>=5.0.0",
    "unstructured>=0.16.0",
    "pillow>=10.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",

    # Storage
    "boto3>=1.35.0",

    # Auth
    "firebase-admin>=6.5.0",

    # Cache
    "redis>=5.0.0",

    # Billing
    "stripe>=11.0.0",

    # Search (for deep research)
    "tavily-python>=0.5.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "eval_type_backport>=0.2.0",  # Python 3.9 support for X | Y syntax.
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
